Move useLocalUnitIds to own file. Remove its use in UnitComparisonView.
